refactor(launcher): extract X11 toolkit override + display diagnostics
Main.kt was carrying two unrelated Linux/X11 helpers:
  - setLinuxXToolkitAppClassName() reflects into sun.awt.X11.XToolkit
    to force WM_CLASS before any window is created.
  - logToolkitAndSession() emits one INFO line tagging the AWT toolkit
    + Wayland/X11 session-type env vars for crash-bundle triage.

Both run before Compose application{} and have no UI concerns, so
they belong in client-launcher. New launcher.bootstrap/ package
becomes the home for pre-Compose JVM-process setup helpers; future
LauncherBootstrap + AutoLoginCoordinator land here too.

Main.kt 822 -> 770 LOC. Call sites updated, unused
java.awt.Toolkit import dropped.
